implement login system with cookie

+++ src/main/java/com/webserver/boundary/ApiResource.java
@@ -1,14 +1,15 @@
 package com.webserver.boundary;
 
+import com.webserver.config.CookieConfigure;
 import com.webserver.model.Task;
 import com.webserver.services.data.TaskRepository;
+import com.webserver.services.data.UserCookies;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @Controller
 @RequestMapping("/api")
@@ -16,22 +17,30 @@ public class ApiResource {
     @Autowired
     TaskRepository taskRepository;
 
+    UserCookies userCookies = new UserCookies();
+
     @RequestMapping("/{todoId}/show")
-    public String auft(@PathVariable("todoId")Long id, Model model){
+    public String auft(@PathVariable("todoId")Long id, Model model, @CookieValue("Login_Hash_Value") Optional<String> cookieHash){
+        if(!userCookies.cookieHash.contains(cookieHash.get())) return "pages/login.html";
+
         Task task = taskRepository.findById(id).get();
         model.addAttribute("todo", task);
         return "pages/todo.html";
     }
 
     @RequestMapping("/{todoId}/edit")
-    public String todoEditieren(@PathVariable("todoId")Long id, Model model){
+    public String todoEditieren(@PathVariable("todoId")Long id, Model model, @CookieValue("Login_Hash_Value") Optional<String> cookieHash){
+        if(!userCookies.cookieHash.contains(cookieHash.get())) return "pages/login.html";
+
         Task task = taskRepository.findById(id).get();
         model.addAttribute("todo", task);
         return "pages/edit.html";
     }
 
     @RequestMapping(value = "/edit",method = RequestMethod.POST)
-    public String editieren(@RequestParam(name = "id") Long id, @RequestParam(name = "title") String title, @RequestParam(name = "inhalt") String inhalt, @RequestParam(name = "status") int status, Model model){
+    public String editieren(@RequestParam(name = "id") Long id, @RequestParam(name = "title") String title, @RequestParam(name = "inhalt") String inhalt, @RequestParam(name = "status") int status, Model model, @CookieValue("Login_Hash_Value") Optional<String> cookieHash){
+        if(!userCookies.cookieHash.contains(cookieHash.get())) return "pages/login.html";
+
         if(id >= 0 && title.length()>=3 && inhalt.length()>=3 && status >= 0) {
             Task task = taskRepository.findById(id).get();
             task.setTitle(title);
@@ -49,7 +58,9 @@ public class ApiResource {
     }
 
     @RequestMapping("/{todoId}/delete")
-    public String editieren(@PathVariable("todoId") Long id, Model model){
+    public String editieren(@PathVariable("todoId") Long id, Model model, @CookieValue("Login_Hash_Value") Optional<String> cookieHash){
+        if(!userCookies.cookieHash.contains(cookieHash.get())) return "pages/login.html";
+
         taskRepository.deleteById(id);
 
         model.addAttribute("allToDos", taskRepository.findAll());
@@ -58,7 +69,9 @@ public class ApiResource {
     }
 
     @RequestMapping(value = "/add",method = RequestMethod.POST)
-    public String addTodo(@RequestParam("title") String title, @RequestParam("inhalt") String inhalt, Model model){
+    public String addTodo(@RequestParam("title") String title, @RequestParam("inhalt") String inhalt, Model model, @CookieValue("Login_Hash_Value") Optional<String> cookieHash){
+        if(!userCookies.cookieHash.contains(cookieHash.get())) return "pages/login.html";
+
         if(title.length()>=3 && inhalt.length()>=3) {
             Task task = new Task(title,inhalt,0);
             taskRepository.save(task);
